video element before styling

+++ src/controllers/videoController.js
@@ -1,8 +1,11 @@
 import Video from "../models/Video";
-
+import User from "../models/User";
 export const home = async (req, res) => {
   try {
-    const videos = await Video.find({}).sort({ createdAt: "desc" });
+    const videos = await await Video.find({})
+      .sort({ createdAt: "desc" })
+      .populate("owner");
+
     return res.render("home", { pageTitle: "Home", videos });
   } catch (error) {
     return res.send(error);
@@ -10,7 +13,7 @@ export const home = async (req, res) => {
 };
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
     return res.render("404", { pageTitle: "Video not found." });
   }
@@ -21,8 +24,17 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
-
+  console.log(video.owner, _id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", {
     pageTitle: `Editing: ${video.title} `,
     video,
@@ -30,10 +42,16 @@ export const getEdit = async (req, res) => {
 };
 export const postEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -50,26 +68,45 @@ export const getUpload = (req, res) => {
 };
 export const postUpload = async (req, res) => {
   // here we will add  a video to the videos array.
+  const {
+    user: { _id },
+  } = req.session;
+  const { file } = req;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl: file.path,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
 
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
+    console.log(error);
     return res.status(404).render("upload", {
-      pageTitle: "Upload Video",
+      pageTitle: "Upload Videos",
       errorMessage: error._message,
     });
   }
 };
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found." });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
-
   //delete video
   return res.redirect("/");
 };
@@ -81,7 +118,7 @@ export const search = async (req, res) => {
       title: {
         $regex: new RegExp(keyword, "i"),
       },
-    });
+    }).populate("owner");
     return res.render("search", { pageTitle: "Search", videos });
   }
   return res.render("search", { pageTitle: "Search", videos: [] });
